requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
faiss-cpu
joblib
numpy